web/pyproject.toml
[tool.poetry]
name = "cost-manager"
version = "0.1.0"
description = ""
authors = ["MattheusBy <mattaudia6@gmail.com>"]
readme = "README.md"
packages = [{include = "cost_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
drf-registration = "^0.1.3"
postgres = "^4.0"
django-filter = "^22.1"
celery = "^5.2.7"
redis = "^4.3.4"
django-celery-beat = "^2.4.0"
export = "^0.2.0"
django-celery-results = "^2.4.0"
shared = "^0.0.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
